Explicitly link libplist with libupshared such that upowerd builds
I wish this was a less common problem, but simply put, upstream only
cared about Requires and not Requires.private in pkg-config files for
the libraries they depend on. In this case, libimobiledevice depends
on libplist and so does the libimobiledevice support in upower - it
works upstream because -lplist is in Requires by default.

Ubuntu patches or upstream updates move -lplist into Requires.private
therefore autotools do not pick it up on PKG_CHECK_MODULES and using
--as-needed linker functionality makes this all the worse.

The problem is that just because a dependent library uses functionality
you CANNOT rely on that dependent library having all IT'S dependent
libraries linking to your application, otherwise every library in the
system might end up linked to every other. In this case we have the
source src/linux/up-device-idevice.c doing #include <plist/plist.h> and then
happily using plist functions.

Using a library's functions means we need to explicitly check and depend
on it and not assume that some other library will pull in dependencies.

Signed-off-by: Matt Sealey <matt@genesi-usa.com>