build with newer version of gobject-introspection
There may be a case for actually build-depending on gobject-introspection
but we don't enable it.

Simply, it was using some weirdo GNU extension and newer autotools (2.67
vs. 2.66) seems to have a real problem with this without the correct flags.
gobject-introspection support has been updated to work around this behavior.

This patch was derived from one upstream (0.9.6) which was broken..

Signed-off-by: Matt Sealey <matt@genesi-usa.com>